requirements.txt
django==3.2.6
django-environ==0.4.5
pillow==8.2.0
stripe==2.60.0
django-allauth==0.45.0
django-tailwind==2.2.0
django-crispy-forms==1.12.0
crispy-tailwind==0.5.0
psycopg2==2.9.1
djangorestframework==3.12.4